+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All major and minor version changes will be documented in this file. Details of patch-level version changes can be found in [commit messages](https://github.com/gchq/CyberChef/commits/master).
 
 
+### [8.36.0] - 2019-07-03
+- 'PGP Verify' operation added [@artemisbot] | [#585]
+
+### [8.35.0] - 2019-07-03
+- 'Sharpen Image', 'Convert Image Format' and 'Add Text To Image' operations added [@j433866] | [#515]
+
 ### [8.34.0] - 2019-06-28
 - Various new visualisations added to the 'Entropy' operation [@MShwed] | [#535]
 - Efficiency improvements made to the 'Entropy' operation for large file support [@n1474335]
@@ -140,6 +146,8 @@ All major and minor version changes will be documented in this file. Details of 
 
 
 
+[8.36.0]: https://github.com/gchq/CyberChef/releases/tag/v8.36.0
+[8.35.0]: https://github.com/gchq/CyberChef/releases/tag/v8.35.0
 [8.34.0]: https://github.com/gchq/CyberChef/releases/tag/v8.34.0
 [8.33.0]: https://github.com/gchq/CyberChef/releases/tag/v8.33.0
 [8.32.0]: https://github.com/gchq/CyberChef/releases/tag/v8.32.0
@@ -244,9 +252,11 @@ All major and minor version changes will be documented in this file. Details of 
 [#489]: https://github.com/gchq/CyberChef/pull/489
 [#496]: https://github.com/gchq/CyberChef/pull/496
 [#506]: https://github.com/gchq/CyberChef/pull/506
+[#515]: https://github.com/gchq/CyberChef/pull/515
 [#516]: https://github.com/gchq/CyberChef/pull/516
 [#525]: https://github.com/gchq/CyberChef/pull/525
 [#531]: https://github.com/gchq/CyberChef/pull/531
 [#533]: https://github.com/gchq/CyberChef/pull/533
 [#535]: https://github.com/gchq/CyberChef/pull/535
 [#571]: https://github.com/gchq/CyberChef/pull/571
+[#585]: https://github.com/gchq/CyberChef/pull/585